+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Pinned the dependency `joblib` at `< 1.3.0`, since `joblib >= 1.3.0` is incompatible
   with the current civis-python codebase. (#469)
 - Changed `civis.io.civis_file_to_table` to not rely on table ids for determining a table's existence (#470)
+- Broke out the "API Resources" documentation page into individual endpoint pages (#471)
 
 ### Deprecated
 ### Removed